separate OpenPinDev into its own module, for stubbing on platforms without hidapi library support
