+++ make/CMakeLists_gl-win-x86.txt
@@ -74,6 +74,7 @@ add_executable(vpinball WIN32
    src/utils/eventproxy.h
    src/parts/backGlass.cpp
    src/core/pininput.cpp
+   src/core/pininput_OpenPinDev.cpp
    src/parts/pinbinary.cpp
    src/core/vpversion.h
