+++ standalone/cmake/CMakeLists_bgfx-android-arm64-v8a.txt
@@ -50,6 +50,7 @@ add_library(vpinball SHARED
    src/utils/eventproxy.h
    src/parts/backGlass.cpp
    src/core/pininput.cpp
+   src/core/pininput_OpenPinDev_stub.cpp
    src/parts/pinbinary.cpp
    src/core/vpversion.h
 
@@ -686,7 +687,6 @@ target_link_libraries(vpinball
    SDL2_ttf
    FreeImage
    bass
-   hidapi
    pinmame.3.6
    altsound
    dmdutil
