+++ standalone/cmake/CMakeLists_bgfx-ios-arm64.txt
@@ -55,6 +55,7 @@ add_executable(vpinball
    src/utils/eventproxy.h
    src/parts/backGlass.cpp
    src/core/pininput.cpp
+   src/core/pininput_OpenPinDev_stub.cpp
    src/parts/pinbinary.cpp
    src/core/vpversion.h
 
@@ -707,7 +708,6 @@ target_link_libraries(vpinball PUBLIC
    "-lSDL2_ttf"
    "-lfreeimage"
    "-lbass"
-   "-lhidapi"
    "-lpinmame.3.6"
    "-laltsound"
    "-lzedmd"
