+++ standalone/cmake/CMakeLists_bgfx-linux-aarch64.txt
@@ -63,6 +63,7 @@ add_executable(vpinball
    src/utils/eventproxy.h
    src/parts/backGlass.cpp
    src/core/pininput.cpp
+   src/core/pininput_OpenPinDev.cpp
    src/parts/pinbinary.cpp
    src/core/vpversion.h
 
@@ -694,7 +695,7 @@ target_link_libraries(vpinball PUBLIC
    SDL2_ttf
    freeimage
    bass
-   hidapi
+   hidapi-hidraw
    -l:libpinmame.so.3.6
    altsound
    dmdutil
