+++ standalone/cmake/CMakeLists_gl-macos-x64.txt
@@ -59,6 +59,7 @@ add_executable(vpinball
    src/utils/eventproxy.h
    src/parts/backGlass.cpp
    src/core/pininput.cpp
+   src/core/pininput_OpenPinDev.cpp
    src/parts/pinbinary.cpp
    src/core/vpversion.h
